Add landing page
Contribute to #46
