+++ app/assets/scripts/components/global/data-browser/bisector.layer.js
