+++ app/assets/scripts/components/global/data-browser/chart.js
