+++ app/assets/scripts/components/global/data-browser/data-extent.layer.js
