+++ app/assets/scripts/components/global/data-browser/data-points.layer.js
