+++ app/assets/scripts/components/global/data-browser/debug.layer.js
