+++ app/assets/scripts/components/global/data-browser/popover.js
