+++ app/assets/scripts/components/global/data-browser/xaxis.layer.js
