+++ app/assets/scripts/components/global/data-browser/yaxis.layer.js
