+++ app/assets/scripts/components/global/data-layers-block.js
