+++ app/assets/scripts/components/global/date-picker.js
