+++ app/assets/scripts/components/global/filter-aoi.js
