+++ app/assets/scripts/components/global/layer.js
