+++ app/assets/scripts/components/global/map-message.js
