+++ app/assets/scripts/components/global/mb-aoi-draw.js
