+++ app/assets/scripts/components/global/mgd-select-rectangle-mode.js
