+++ app/assets/scripts/components/global/prime-panel.js
