+++ app/assets/scripts/components/global/sec-panel.js
