+++ app/assets/scripts/components/global/timeline.js
